backend/martinbullman/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
astroid==3.3.5
certifi==2024.8.30
charset-normalizer==3.4.0
decouple==0.0.7
dill==0.3.9
distro==1.9.0
Django==5.1.4
django-cors-headers==4.5.0
djangorestframework==3.15.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
isort==5.13.2
jiter==0.6.1
mailjet-rest==1.3.4
mccabe==0.7.0
Naked==0.1.32
openai==1.52.2
packaging==24.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg==3.2.3
psycopg2-binary==2.9.10
pydantic==2.6.3
pydantic_core==2.16.3
pylint==3.3.1
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
pytest==8.3.4
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.1
status==0.2.5
tomlkit==0.13.2
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
utils==1.0.2
